all instructions in disasm done, bitches

+++ c_impl/include/chip8_cpu.h
@@ -6,7 +6,7 @@ typedef struct chip8_reg_t {
 	uint8_t  V[16];
 	uint16_t I;
 	uint16_t PC;
-	uint8_t  SP;
+	uint16_t SP;
 	uint8_t  DT;
 	uint8_t  ST;
 } chip8_reg_t;
